pbl_admin/requirements.txt
# pbl_admin/requirements.txt
streamlit
requests
python-dotenv